dom/media/webaudio/test/test_bug866570.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Crashtest for bug 859600</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

(new AudioContext()).foo = null;
ok(true, "The test should not fatally assert");

</script>
</pre>
</body>
</html>
